Config CustomerServlet doDelete method by using http status codes

+++ src/lk/ijse/dep/web/api/CustomerServlet.java
@@ -164,27 +164,56 @@ public class CustomerServlet extends HttpServlet {
 
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+        /* CORS policy */
+        resp.setHeader("Access-Control-Allow-Origin", CommonConstants.FRONTEND_URL);
+
         String id = req.getParameter("id");
+
+        /* Validation */
+        if (id == null || id.trim().isEmpty() || !id.trim().matches("\\d+")) {
+            resp.sendError(HttpServletResponse.SC_BAD_REQUEST);
+            return;
+        }
 
         BasicDataSource cp = (BasicDataSource) getServletContext().getAttribute("cp");
 
         try {
             Class.forName(CommonConstants.MYSQL_DRIVER_CLASS_NAME);
-            Connection connection = cp.getConnection();
-            PreparedStatement pstm = connection.prepareStatement("DELETE FROM customer WHERE `id` = ?");
+            try (Connection connection = cp.getConnection()) {
 
-            pstm.setObject(1, id);
-            boolean success = pstm.executeUpdate() > 0;
-            if (success) {
-                /* deleted successfully ---> 204 status code */
-                resp.setStatus(HttpServletResponse.SC_NO_CONTENT);
-            } else {
+                /* Check if there is a record for the given ID */
+                PreparedStatement preparedStatement = connection.prepareStatement("SELECT * FROM customer WHERE id=?");
+                preparedStatement.setObject(1, id);
+
+                /* if customer does not exists for that ID, send NOT_FOUND error */
+                if (!(preparedStatement.executeQuery().next())) {
+                    resp.sendError(HttpServletResponse.SC_NOT_FOUND);
+                    return;
+                }
+
+                /* Database operation - delete customer record */
+                PreparedStatement pstm = connection.prepareStatement("DELETE FROM customer WHERE `id` = ?");
+
+                pstm.setObject(1, id);
+                boolean success = pstm.executeUpdate() > 0;
+                if (success) {
+                    /* deleted successfully ---> 204 status code */
+                    resp.setStatus(HttpServletResponse.SC_NO_CONTENT);
+                } else {
+                    resp.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
+                }
+
+            }catch (SQLIntegrityConstraintViolationException throwables) {
+                throwables.printStackTrace();
+                resp.setStatus(HttpServletResponse.SC_BAD_REQUEST);
+            } catch (SQLException throwables) {
+                throwables.printStackTrace();
                 resp.setStatus(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
             }
-
-            connection.close();
-        } catch (SQLException | ClassNotFoundException throwables) {
+        } catch (ClassNotFoundException throwables) {
             throwables.printStackTrace();
+            resp.sendError(HttpServletResponse.SC_INTERNAL_SERVER_ERROR);
         }
 
     }//doDelete
@@ -237,7 +266,7 @@ public class CustomerServlet extends HttpServlet {
             preparedStatement.setObject(1, id);
 
             /* if customer does not exists for that ID */
-            if (!(preparedStatement.executeQuery().next())){
+            if (!(preparedStatement.executeQuery().next())) {
                 resp.sendError(HttpServletResponse.SC_NOT_FOUND);
                 return;
             }
